+++ lib/fastpix_client/models/components/playlistbyidresponsemedialistitem.rb
@@ -15,7 +15,7 @@ module FastpixClient
         # Timestamp of media creation in the workspace.
         field :created_at, Crystalline::Nilable.new(::DateTime), { 'format_json': { 'letter_case': ::FastpixClient::Utils.field_name('createdAt'), 'decoder': Utils.datetime_from_iso_format(true) } }
         # Duration of the media in hh:mm:ss format.
-        field :duration, Crystalline::Nilable.new(::String), { 'format_json': { 'letter_case': ::FastpixClient::Utils.field_name('duration') } }
+        field :duration, Crystalline::Nilable.new(::Float), { 'format_json': { 'letter_case': ::FastpixClient::Utils.field_name('duration') } }
         # unique id of the particular media.
         field :id, Crystalline::Nilable.new(::String), { 'format_json': { 'letter_case': ::FastpixClient::Utils.field_name('id') } }
         # source resolution of the media.
@@ -29,7 +29,7 @@ module FastpixClient
         # Title of the media.
         field :title, Crystalline::Nilable.new(::String), { 'format_json': { 'letter_case': ::FastpixClient::Utils.field_name('title') } }
 
-        sig { params(created_at: T.nilable(::DateTime), duration: T.nilable(::String), id: T.nilable(::String), source_resolution: T.nilable(::String), status: T.nilable(::String), thumbnail: T.nilable(::String), creator_id: T.nilable(::String), title: T.nilable(::String)).void }
+        sig { params(created_at: T.nilable(::DateTime), duration: T.nilable(::Float), id: T.nilable(::String), source_resolution: T.nilable(::String), status: T.nilable(::String), thumbnail: T.nilable(::String), creator_id: T.nilable(::String), title: T.nilable(::String)).void }
         def initialize(created_at: nil, duration: nil, id: nil, source_resolution: nil, status: nil, thumbnail: nil, creator_id: nil, title: nil)
           @created_at = created_at
           @duration = duration